custom-images/uidgid/requirements.txt
pyyaml
requests
filelock
text-unidecode
unidecode
flask
gunicorn
quart
#quart-schema
hypercorn
